refactor(framework): merge VuetifyUseOptions and VuetifyPreset
clean-up types and process
