+++ packages/vuetify/src/index.ts
@@ -1,6 +1,6 @@
 // Core
 import { App } from 'vue'
-import { VuetifyUseOptions } from 'types'
+import { GlobalVuetifyPreset } from 'types'
 import { install as VuetifyInstall } from './install'
 
 // Functionality
@@ -9,12 +9,13 @@ import * as directives from './directives'
 
 // Wrapped install function to register all components by default
 // original install for a-la-carte is re-exported in entry-lib.ts
-export default function install (app: App, args: VuetifyUseOptions) {
-  VuetifyInstall(app, {
+export default function install (app: App, preset: GlobalVuetifyPreset = {}) {
+  const args = Object.assign({
     components,
     directives,
-    ...args,
-  })
+  }, preset)
+
+  VuetifyInstall(app, args)
 }
 
 export { useVuetify, version } from './framework'
