--- packages/vuetify/types/services/presets.d.ts
@@ -1,62 +1,0 @@
-// Services
-import {
-  Breakpoint,
-  BreakpointOptions,
-} from 'vuetify/types/services/breakpoint'
-import {
-  Icons,
-  IconsOptions,
-} from 'vuetify/types/services/icons'
-import {
-  Lang,
-  LangOptions,
-} from 'vuetify/types/services/lang'
-import {
-  Theme,
-  ThemeOptions,
-} from 'vuetify/types/services/theme'
-
-export interface VuetifyPreset {
-  breakpoint: {
-    scrollBarWidth: Breakpoint['scrollBarWidth']
-    thresholds: Breakpoint['thresholds']
-  }
-  icons: {
-    iconfont: Icons['iconfont']
-    // TODO: Remove partial for v3
-    values: Partial<Icons['values']>
-  }
-  lang: {
-    current: Lang['current']
-    locales: Lang['locales']
-    t: Lang['t']
-  }
-  theme: {
-    dark: Theme['dark']
-    default: Theme['default']
-    disable: Theme['disable']
-    options: Theme['options']
-    themes: Theme['themes']
-  }
-
-  [name: string]: any
-}
-
-export interface UserVuetifyPreset extends GlobalVuetifyPreset {
-  preset?: GlobalVuetifyPreset
-
-  [name: string]: any
-}
-
-export interface GlobalVuetifyPreset {
-  breakpoint?: BreakpointOptions
-  icons?: IconsOptions
-  lang?: LangOptions
-  theme?: ThemeOptions
-
-  [name: string]: any
-}
-
-export interface Presets {
-  preset: UserVuetifyPreset
-}
